src/main/java/guru/springfamework/api/v1/mapper/CategoryMapperImpl.java
package guru.springfamework.api.v1.mapper;

import guru.springfamework.api.v1.model.CategoryDTO;
import guru.springfamework.domain.Category;
import org.springframework.stereotype.Component;

@Component
public class CategoryMapperImpl   {

    public CategoryDTO categoryToCategoryDTO(Category category) {
        if (category == null) {
            return null;
        }
        CategoryDTO categoryDTO = new CategoryDTO();
        categoryDTO.setId(category.getId());
        categoryDTO.setName(category.getName());
        return categoryDTO;
    }

}
